requirements.txt
crewai
langchain-groq
streamlit
python-dotenv
yfinance
duckduckgo-search
feedparser
llama-parse